openblas -- try to add test routine
