+++ rstsr-blas-traits/driver_impl/lapacke/eigh/syevd.rs
@@ -58,26 +58,3 @@ impl SYEVDDriverAPI<T> for DeviceBLAS {
         )
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::DeviceBLAS;
-    use rstsr_core::prelude_dev::*;
-    use rstsr_test_manifest::get_vec;
-
-    #[test]
-    fn playground() {
-        let device = DeviceBLAS::default();
-        let la = [2048, 2048].c();
-        let a = Tensor::new(Storage::new(get_vec::<f64>('a').into(), device.clone()), la);
-        let driver = DSYEVD::default().a(a.view()).build().unwrap();
-        let (c, w) = driver.run().unwrap();
-        let c = c.into_owned();
-        // println!("{:?}", c);
-        println!("{:?}", w);
-        assert!(c.c_contig());
-        println!("{:?}", fingerprint(&c));
-        // assert!((fingerprint(&c) - -8345.684144788995).abs() < 1e-8);
-    }
-}
