+++ rstsr-openblas/tests/test_driver_impl/mod.rs
@@ -1,0 +1,1 @@
+mod lapack_eigh;
